+++ src/clikEvent.js
@@ -1,0 +1,12 @@
+import { Render } from './render.js';
+
+const clickEvent = (home) => {
+    home = new Render('p',content).classElement('para');
+    home.textContent = `EATING, ELEVATED Discover an elevated dining experience in downtown Portland – a Pan Asian 
+        kitchen from Sage Restaurant Group. Floating atop the landmark Meier & 
+        Frank Building, Departure’s ambitious menu captivates diners with its 
+        masterful remix of authentic coastal cuisines.`
+
+}
+
+export { clickEvent };
