+++ src/style.scss
@@ -24,7 +24,7 @@ html{
       align-items: center;
       color: black;
       li{
-        width: 80%;
+        width: 100%;
         height: 100%;
         display: flex;
         align-items: center;
@@ -41,4 +41,14 @@ html{
       }
     }
   }
+  .para{
+    position: relative;
+    color: white;
+    bottom: 600px;
+    font-size: 1.2em;
+    line-height: 2em;
+    padding: 150px;
+  }
 }
+
+
